Test for large_num_segments is parametrized by three modulus params
The test cooks input sequence in the form of F(i+1) - F(i) for
0 <= i < num_segments * segment_size.

Fixed-size segmented reduction is expected to to be

[ F(segment_size * (k + 1)) - F(segment_size * k) for
     k in range(num_segments) ]

All operations (addition (+), inversion (-)) are assumed to take
place in a ring.

In this test the ring is a cycling ring of integers modulo 7.
F(i) = g(i) * g(i+1) % 7, where g: i -> (i % 5 + i % 3).

Note that `my_add` computes addition modulo 7 too.

Choice of moduli 3, 5, 7 is arbitrary and may be changed. The test
is expected to pass so long as they are changed consistently throughout
the test.