Add test_large_num_segments3
This example computes segmented reduce of input constant iterator
of int8(1) values. The offsets iterators are crafted to that segment
lengths are

1, 2, ..., (p - 2), (p - 1), 1, 2, ...,  (p - 2), (p - 1), 1, ...

The expected sums would equal these lengths.

The example is run for (2**15 + 2**3) * 2**16 number of segments
which is INT_MAX and about 2**19 elements.

1. Input iterator changed to have type int16
2. Initial value changed to have type int16
3. Changed range of change of segments from [1, p] to
   [m0, m0 + p) and set m0, p as they are set in native test.

   Specifically, m0, p = 265, 163

This means segment sizes vary from 265 elements to 428 elements,
which is an improvement over previous situation where it was varying
from 1 to 113 elements (and most threads in a 256-string block were
unused).